@@ -8,7 +8,7 @@
 
 package com.akingyin.base.net.mode;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import java.util.HashMap;
 import java.util.Map;
 import okhttp3.Request;
