@@ -8,8 +8,8 @@
 
 package com.akingyin.base.net.okhttp;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.text.TextUtils;
 import android.util.Log;
 import com.akingyin.base.net.okhttp.Interceptor.LoginStatusInterceptor;
