@@ -1,9 +1,6 @@
 package com.akingyin.base
 
 import android.annotation.SuppressLint
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentPagerAdapter
 import java.util.*
 
 /**
@@ -16,38 +13,38 @@ import java.util.*
  * 应该使用FragmentStatePagerAdapter。
  * @version V1.0
  */
-class BaseFragmentAdapter : FragmentPagerAdapter{
+class BaseFragmentAdapter : androidx.fragment.app.FragmentPagerAdapter {
 
 
-    private   var  fragmentlist:List<Fragment>?= ArrayList()
+    private   var  fragmentlist:List<androidx.fragment.app.Fragment>?= ArrayList()
 
     private   var  mTitles:List<String>?=null
 
-    constructor(fm: FragmentManager?, fragmentlist: List<Fragment>?) : super(fm) {
+    constructor(fm: androidx.fragment.app.FragmentManager?, fragmentlist: List<androidx.fragment.app.Fragment>?) : super(fm) {
         this.fragmentlist = fragmentlist
     }
 
-    constructor(fm: FragmentManager?, fragmentlist: List<Fragment>?, mTitles: List<String>?) : super(fm) {
+    constructor(fm: androidx.fragment.app.FragmentManager?, fragmentlist: List<androidx.fragment.app.Fragment>?, mTitles: List<String>?) : super(fm) {
         this.fragmentlist = fragmentlist
         this.mTitles = mTitles
     }
 
     //刷新fragment
     @SuppressLint("CommitTransaction")
-    private fun setFragments(fm: FragmentManager, fragments: List<Fragment>, mTitles: List<String>) {
+    private fun setFragments(fm: androidx.fragment.app.FragmentManager, fragments: List<androidx.fragment.app.Fragment>, mTitles: List<String>) {
         this.mTitles = mTitles
         if (this.fragmentlist != null) {
             val ft = fm.beginTransaction()
             fragmentlist?.forEach {
                 ft.remove(it)
             }
-            ft?.commitAllowingStateLoss()
+            ft.commitAllowingStateLoss()
             fm.executePendingTransactions()
         }
         this.fragmentlist = fragments
         notifyDataSetChanged()
     }
-    override fun getItem(p0: Int): Fragment {
+    override fun getItem(p0: Int): androidx.fragment.app.Fragment {
         return  fragmentlist!![p0]
     }
 
